rename image_plane contsub tools script
